Create Code of Hammurabi.md
Originally published on 20-Oct 2021. I think this is a bit much and wouldn't make it into modern regulations for medical ethics:

"If a physician make a deep incision upon a man with his bronze lancet and cause the man's death, or operate on the eye socket of a man with a bronze lancet and destroy the man's eye, they shall cut off his hand" (Smith, [1931](https://www.lawbookexchange.com/pages/books/40721/j-m-powis-smith/the-origin-and-history-of-hebrew-law): 211).