Create Novum Organum by Francis Bacon.md
Legacy story [published](https://www.tiki-toki.com/timeline/entry/1753034/A-History-of-Research-Ethics/#vars!panel=16443473!) on 16 or 17-Oct.

Like with Galilei's, this is a case that has been scrutinised *ad nauseam*. Will require many sources for a useful contribution. The link to research ethics boils down to how we conduct science, how we define scientific-vs.-unscientific "knowledge," and what we value (is "dominion over nature possibly construed as a bit bad for the environment, maybe?).

A picture (media) would be nice too.