Create Hind Swaraj, by Mahatma Gandhi.md
Legacy story [published](https://www.tiki-toki.com/timeline/entry/1753034/A-History-of-Research-Ethics/#vars!panel=16443706!) 16 or 17-Oct 2021.

This story fits within the wider history of copyright, which The Timeline incorrectly sees as beginning with the Berne Convention of 1886.

This story also hints at the challenge found with the "Principle of Patient Autonomy" story; i.e.: it draws heavily on one paper, so should it link to that paper? Conversely, should it link to the site of the book's current publisher? If the latter, the idea of linking directly to papers loses strength. If the former, then this story must draw much more on Balganesh.