Create Atomic Bomb, Part I.md
Legacy story [published](https://www.tiki-toki.com/timeline/entry/1753034/A-History-of-Research-Ethics/#vars!panel=16443826!) 16 or 17-Oct 2021.

The story literally lacks any content. It might be worth describing the recent discovery of nuclear fission. I think this links, for now, only with Rotblat's story and the bombings themselves in August 1945. Still, a link can be drawn to the influence of the eugenics movement on Nazi experimentation.