Create Animal Liberation, by Peter Singer.md
Legacy story [published](https://www.tiki-toki.com/timeline/entry/1753034/A-History-of-Research-Ethics/#vars!panel=16444353!) 16 or 17-Oct 2021.

This is only one of the publications mentioned in [The Principles of Humane Experimental Technique, by Russell and Burch](https://www.tiki-toki.com/timeline/entry/1753034/A-History-of-Research-Ethics/#vars!date=1959-01-01_00:00:00!). Thought needs to be put into whether that story remains as it is (possibly meaning this one can be dropped) or whether this one is expanded.