Create Kimberlé Crenshaw coins "intersectionality".md
Legacy article [published](https://www.tiki-toki.com/timeline/entry/1753034/A-History-of-Research-Ethics/#vars!date=1989-01-01_00:00:00!) 16 or 17-Oct 2021.

This is the first story about the coining of a term I wrote. I think it is important to acknowledge the history of some terms that have come into the mainstream and, being the victim of their own success, have fallen into over- and misuse, probably mostly by defenders of the *status quo*.

I think "intersectionality" is an obvious candidate for a list of terms that have shaped how we conduct ourselves in research and what research is conducted. A broad (GitHub) issue might be worth writing to track relevant terms. (I have thought of Bentham's "panopticon and Anscombe's "consequentialism;" because terms needn't always be invented by their supporters!).