Create Science as Social Knowledge, by Helen Longino.md
Legacy story[published](https://www.tiki-toki.com/timeline/entry/1753034/A-History-of-Research-Ethics/#vars!panel=16444411!) 16 or 17-Oct 2021.

Do I want The Timeline to provide examples of things philosophers have said that are valuable to research ethics? Yes. Do I need to share Longino's particular view? Probably not, but I like it.

Nonetheless, there is a history of idealism, empiricism, logical positivism and all that jazz that this story asks for in the wider timeline.

Another thematic (Github) issue might be on PHILOSOPHY OF SCIENCE.