Create CE 1782 Princess Dashkova.md
As per #19 :

### Elevator Pitch
Might she be the first woman to found a modern science academy? In 1782, she was appointed to direct the Petersburg Academy of Arts and Sciences, and then became the first president of the Russian Academy (The Editors of the Encyclopaedia Britannica, [n.d.](https://www.britannica.com/biography/Yekaterina-Romanovna-Vorontsova-Knyaginya-Dashkova)). Later, in 1789, she became the first female member of both the German Leopoldina (Leopoldina.org, [n.d.](https://www.leopoldina.org/en/about-us/about-the-leopoldina/history/the-history-of-the-leopoldina/the-academys-nomadic-years/)) and the American Philosophical Society (Woronzoff-Dashkoff, [1996](https://www.jstor.org/stable/987316)).
### Justification
I am not seeing references to her being a scientist, so her joining science academies might be a sign of increased interest on the part of these academies to engage with "policy makers" and/or non-scientists ("lay persons," if you will).
### Unformed Thoughts
Still a stab in the dark, I only found her name from writing on the Leopoldina. I might need to go through the other three scientific institutes to see when they first had women join.