Update BCE 1755 Code of Hammurabi.md
The migration shall now pause!

@yochannah, I'll pause at this story because (i) I will need to see pull requests merged to make sure it's all running smoothly, and (ii) I think there is a missing button (which is entirely on me!).

Some stories have a "Find out more" button on Tiki-Toki, like [this one](https://www.tiki-toki.com/timeline/entry/1753034/A-History-of-Research-Ethics/#vars!date=3004_BC-09-28_22:35:50!). This is what the link in [line 27 of the stories template](https://github.com/Ismael-KG/A-History-of-Research-Ethics/blob/main/A%20Template%20for%20Stories.md?plain=1) would include (or "N/A").

Do you think you can add something to the yaml (is this the term?) to show up on the site? 🍪